Rewrite random word selection query
`ORDER BY RAND()` is incredibly slow on larger tables. `SELECT COUNT(*)`
followed by an `OFFSET` is much faster. This doesn't make much of a
difference for the 850-word Basic English list, but is a good change to
make anyway, especially if the wordlist is to be expanded in the future.